Use unique instance id as runner name, to prevent name clash
Sometimes a GitHub actions job does not cleanup the self-hosted runner
from the list of registered runners. Then a runner with a ip address
as the name is left.

A new runner that does potentially get the same IP as the previous runner
is now no longer able to connect to the GitHub runner API as the runner
is already registered (but offline).

Using the EC2 instance id does ensure that the name is somewhat unique and
does not interfere with left over, not cleaned up offline runners.

KNUTH-62626